Переносит функцию получения пака id в utils
